int/git: inits to enable testing of git related logic
